[REFACTOR] : ♻️ Modal 👉 CustomModal
